Shooting architecture draft.
