+++ game/src/main/java/chon/group/game/domain/agent/Agent.java
@@ -3,54 +3,27 @@ package chon.group.game.domain.agent;
 import java.util.ArrayList;
 import java.util.List;
 
+import chon.group.game.core.Entity;
 import chon.group.game.messaging.Message;
-import javafx.scene.SnapshotParameters;
-import javafx.scene.image.Image;
-import javafx.scene.image.ImageView;
-import javafx.scene.paint.Color;
 
 /**
  * Represents an agent in the game, with properties such as position, size,
  * speed, and image.
  * The agent can move in specific directions and chase a target.
  */
-public class Agent {
-
-    /** X position (horizontal) of the agent. */
-    private int posX;
-
-    /** Y (vertical) position of the agent. */
-    private int posY;
-
-    /** Height of the agent. */
-    private int height;
-
-    /** Width of the agent. */
-    private int width;
-
-    /** Agent speed. */
-    private int speed;
-
-    /** Image representing the agent. */
-    private Image image;
-
-    /** Indicates if the agent is facing left. */
-    private boolean flipped = false;
-
-    /** The initial agent's health. */
-    private int health;
-
-    /** The maximum agent's health. */
-    private int fullHealth;
+public class Agent extends Entity {
 
     /* The time of the last hit taken. */
     private long lastHitTime = 0;
 
     /* Flag to control the invulnerability status of the agent. */
-    private boolean invulnerable;
+    private boolean invulnerable = false;
 
     /* Invulnerability (in milliseconds) */
     private final long INVULNERABILITY_COOLDOWN = 500;
+
+    /* The Agent's Weapon */
+    private Weapon weapon;
 
     /**
      * Constructor to initialize the agent properties.
@@ -64,16 +37,7 @@ public class Agent {
      * @param pathImage the path to the agent's image
      */
     public Agent(int posX, int posY, int height, int width, int speed, int health, String pathImage) {
-        this.posX = posX;
-        this.posY = posY;
-        this.height = height;
-        this.width = width;
-        this.speed = speed;
-        this.health = health;
-        this.fullHealth = health;
-        this.image = new Image(getClass().getResource(pathImage).toExternalForm());
-        this.lastHitTime = 0;
-        this.invulnerable = false;
+        super(posX, posY, height, width, speed, health, pathImage);
     }
 
     /**
@@ -89,177 +53,7 @@ public class Agent {
      * @param flipped   the agent's direction (RIGHT=0 or LEFT=1)
      */
     public Agent(int posX, int posY, int height, int width, int speed, int health, String pathImage, boolean flipped) {
-        this.posX = posX;
-        this.posY = posY;
-        this.height = height;
-        this.width = width;
-        this.speed = speed;
-        this.health = health;
-        this.fullHealth = health;
-        this.image = new Image(getClass().getResource(pathImage).toExternalForm());
-        this.flipped = flipped;
-    }
-
-    /**
-     * Gets the X (horizontal) position of the agent.
-     *
-     * @return the X (horizontal) position of the agent
-     */
-    public int getPosX() {
-        return posX;
-    }
-
-    /**
-     * Sets the agent's X (horizontal) position.
-     *
-     * @param posX the new X (horizontal) position
-     */
-    public void setPosX(int posX) {
-        this.posX = posX;
-    }
-
-    /**
-     * Gets the Y (vertical) position of the agent.
-     *
-     * @return the Y (vertical) position of the agent
-     */
-    public int getPosY() {
-        return posY;
-    }
-
-    /**
-     * Sets the Y (vertical) position of the agent.
-     *
-     * @param posY the new Y (vertical) position
-     */
-    public void setPosY(int posY) {
-        this.posY = posY;
-    }
-
-    /**
-     * Gets the height of the agent.
-     *
-     * @return the height of the agent
-     */
-    public int getHeight() {
-        return height;
-    }
-
-    /**
-     * Sets the height of the agent.
-     *
-     * @param height the new height
-     */
-    public void setHeight(int height) {
-        this.height = height;
-    }
-
-    /**
-     * Gets the width of the agent.
-     *
-     * @return the width of the agent
-     */
-    public int getWidth() {
-        return width;
-    }
-
-    /**
-     * Sets the width of the agent.
-     *
-     * @param width the new width
-     */
-    public void setWidth(int width) {
-        this.width = width;
-    }
-
-    /**
-     * Gets the agent's speed.
-     *
-     * @return the agent's speed
-     */
-    public int getSpeed() {
-        return speed;
-    }
-
-    /**
-     * Sets the speed of the agent.
-     *
-     * @param speed the new speed
-     */
-    public void setSpeed(int speed) {
-        this.speed = speed;
-    }
-
-    /**
-     * Gets the agent's health.
-     *
-     * @return the agent's health
-     */
-    public int getHealth() {
-        return health;
-    }
-
-    /**
-     * Sets the health of the agent.
-     *
-     * @param health the new health
-     */
-    public void setHealth(int health) {
-        this.health = health;
-    }
-
-    /**
-     * Gets the agent's maximum health.
-     *
-     * @return the agent's maximum health
-     */
-    public int getFullHealth() {
-        return fullHealth;
-    }
-
-    /**
-     * Sets the maximum health of the agent.
-     *
-     * @param fullHealth the new maximum health
-     */
-    public void setFullHealth(int fullHealth) {
-        this.fullHealth = fullHealth;
-    }
-
-    /**
-     * Gets the agent image.
-     *
-     * @return the agent image
-     */
-    public Image getImage() {
-        return image;
-    }
-
-    /**
-     * Gets the agent flipped status.
-     *
-     * @param image the new image
-     */
-    public void setImage(Image image) {
-        this.image = image;
-    }
-
-    /**
-     * Gets if the agent is flipped.
-     *
-     * @return if the agent is flipped
-     */
-    public boolean isFlipped() {
-        return flipped;
-    }
-
-    /**
-     * Sets the agent flipped status.
-     *
-     * @param flipped the new flipped status
-     */
-    public void setFlipped(boolean flipped) {
-        this.flipped = flipped;
+        super(posX, posY, height, width, speed, health, pathImage, flipped);
     }
 
     /**
@@ -306,24 +100,30 @@ public class Agent {
     }
 
     /**
+     * Gets the agent's weapon.
+     *
+     * @return its weapon.
+     */
+    public Weapon getWeapon() {
+        return weapon;
+    }
+
+    /**
+     * Sets the agent new weapon.
+     *
+     * @param weapon the new weapon
+     */
+    public void setWeapon(Weapon weapon) {
+        this.weapon = weapon;
+    }
+
+    /**
      * Gets if the agent is dead.
      *
      * @return if the agent is dead
      */
     public boolean isDead() {
-        return (this.health <= 0);
-    }
-
-    /**
-     * Flips the Image horizontally.
-     */
-    private void flipImage() {
-        ImageView flippedImage = new ImageView(image);
-        flippedImage.setScaleX(-1);
-        SnapshotParameters params = new SnapshotParameters();
-        params.setFill(Color.TRANSPARENT);
-        this.flipped = !this.flipped;
-        this.image = flippedImage.snapshot(params, null);
+        return (this.getHealth() <= 0);
     }
 
     /**
@@ -334,17 +134,17 @@ public class Agent {
      */
     public void move(List<String> movements) {
         if (movements.contains("RIGHT")) {
-            if (flipped)
+            if (this.isFlipped())
                 this.flipImage();
-            setPosX(posX += speed);
+            setPosX(this.getPosX() + this.getSpeed());
         } else if (movements.contains("LEFT")) {
-            if (!flipped)
+            if (!this.isFlipped())
                 this.flipImage();
-            setPosX(posX -= speed);
+            setPosX(this.getPosX() - this.getSpeed());
         } else if (movements.contains("UP")) {
-            setPosY(posY -= speed);
+            setPosY(this.getPosY() - this.getSpeed());
         } else if (movements.contains("DOWN")) {
-            setPosY(posY += speed);
+            setPosY(this.getPosY() + this.getSpeed());
         }
     }
 
@@ -355,14 +155,14 @@ public class Agent {
      * @param targetY the target's Y (vertical) position
      */
     public void chase(int targetX, int targetY) {
-        if (targetX > this.posX) {
+        if (targetX > this.getPosX()) {
             this.move(new ArrayList<String>(List.of("RIGHT")));
-        } else if (targetX < this.posX) {
+        } else if (targetX < this.getPosX()) {
             this.move(new ArrayList<String>(List.of("LEFT")));
         }
-        if (targetY > this.posY) {
+        if (targetY > this.getPosY()) {
             this.move(new ArrayList<String>(List.of("DOWN")));
-        } else if (targetY < this.posY) {
+        } else if (targetY < this.getPosY()) {
             this.move(new ArrayList<String>(List.of("UP")));
         }
     }
@@ -373,19 +173,20 @@ public class Agent {
      *
      * @param damage the amount of damage to be applied
      */
+    @Override
     public void takeDamage(int damage, List<Message> messages) {
         this.invulnerable = this.updateInvulnerability();
-        if (!this.invulnerable && this.health > 0) {
+        if (!this.invulnerable && this.getHealth() > 0) {
             /* Decrease health. */
-            this.health = health - damage;
+            this.setHealth(this.getHealth() - damage);
             messages.add(new Message(
                     String.valueOf(damage),
                     this.getPosX(),
                     this.getPosY(),
                     25));
             /* After taking the damage, the health must not be negative. */
-            if (this.health < 0)
-                this.health = 0;
+            if (this.getHealth() < 0)
+                this.setHealth(0);
             else
                 this.lastHitTime = System.currentTimeMillis();
         }
