+++ game/src/main/java/chon/group/game/domain/agent/Weapon.java
@@ -1,0 +1,15 @@
+package chon.group.game.domain.agent;
+
+public class Weapon {
+
+    Shot shot;
+
+    public void attack() {
+
+    }
+
+    public Shot fire() {
+        return this.shot;
+    }
+
+}
